Добавляем формы...

+++ MetalAvaloniaReactive/ViewModels/MainAdminViewModel.cs
@@ -23,37 +23,36 @@ public class MainAdminViewModel : ViewModelBase
     private ObservableCollection<Role> _roles;
     private ObservableCollection<CalculationHistory> _calculationHistories;
     private ObservableCollection<Furnace> _furnaces;
-    public event PropertyChangedEventHandler PropertyChanged;
-    public MainAdminViewModel(MainWindowViewModel mainWindowViewModel)
+    public MainAdminViewModel(MainWindowViewModel mainWindowViewModel, bool isAdmin)
     {
+        _mainWindowViewModel = mainWindowViewModel;
+        IsAdmin = isAdmin;
+        ExitFromApplication = ReactiveCommand.CreateFromTask(async () =>
+        {
+            var messageBox = MessageBox.Avalonia.MessageBoxManager.GetMessageBoxStandardWindow("Предупреждение",
+                "Вы уверены, что хотите выйти?", ButtonEnum.YesNo, Icon.Info);
+            var messageBoxResult = await messageBox.Show();
+            if (messageBoxResult == ButtonResult.Yes)
+            {
+                PreparedLocalStorage.ClearLocalStorage();
+                PreparedLocalStorage.SaveLocalStorage();
+                _mainWindowViewModel.Content = new AuthorizationViewModel(_mainWindowViewModel);
+            }
+        });
+        if (!IsAdmin) return;
         try
         {
             this.WhenAnyValue(x => x.Search)
                 .Throttle(TimeSpan.FromMilliseconds(200))
                 .ObserveOn(RxApp.MainThreadScheduler)
                 .Subscribe(Searching);
-            
-            
             _users = Users = new ObservableCollection<User>(UserImplementation.GetAllUsers().Result.OrderBy(u => u.Id));
             _roles = Roles = new ObservableCollection<Role>(RoleImplementation.GetAllRoles().Result.OrderBy(r => r.Id));
             _furnaces = Furnaces = new ObservableCollection<Furnace>(FurnaceImplementation.GetAllFurnaces().Result);
             _calculationHistories = CalculationHistories = new ObservableCollection<CalculationHistory>(CalculationHistoryImplementation.GetAllHistoryRecords().Result.
                 OrderBy(x => x.CreationDate));
             SearchButtonClick = ReactiveCommand.Create(SearchingItems);
-            _mainWindowViewModel = mainWindowViewModel;
             _selectedTabItem = 0;
-            ExitFromApplication = ReactiveCommand.CreateFromTask(async () =>
-            {
-                var messageBox = MessageBox.Avalonia.MessageBoxManager.GetMessageBoxStandardWindow("Предупреждение",
-                    "Вы уверены, что хотите выйти?", ButtonEnum.YesNo, Icon.Info);
-                var messageBoxResult = await messageBox.Show();
-                if (messageBoxResult == ButtonResult.Yes)
-                {
-                    PreparedLocalStorage.ClearLocalStorage();
-                    PreparedLocalStorage.SaveLocalStorage();
-                    _mainWindowViewModel.Content = new AuthorizationViewModel(_mainWindowViewModel);
-                }
-            });
             AddRecordClick = ReactiveCommand.Create<int>(OpenOneRecordView);
             DeleteRecordClick = ReactiveCommand.CreateFromTask<int>(async (id) => { DeleteRecord(id); });
             UpdateRecordClick = ReactiveCommand.Create<int>(OpenOneRecordView);
@@ -64,7 +63,11 @@ public class MainAdminViewModel : ViewModelBase
                 ("Ошибка", ex.Message, ButtonEnum.Ok, Icon.Error);
             messageBox.Show();
         }
+
+
     }
+
+    public bool IsAdmin { get; }
 
     public string Search
     {
@@ -133,7 +136,7 @@ public class MainAdminViewModel : ViewModelBase
                 MessageBox.Avalonia.MessageBoxManager.GetMessageBoxStandardWindow("Инфо", "Запись успешно удалена \t",
                     ButtonEnum.Ok, Icon.Info);
             messageBoxInfo.Show();
-            _mainWindowViewModel.Content = new MainAdminViewModel(_mainWindowViewModel);
+            _mainWindowViewModel.Content = new MainAdminViewModel(_mainWindowViewModel, true);
         }
         catch (Exception ex)
         {
@@ -168,7 +171,8 @@ public class MainAdminViewModel : ViewModelBase
                     .Where(u => u.Login.Trim().ToLower().Contains(s.Trim().ToLower()) || 
                                 u.Surname.Trim().ToLower().Contains(s.Trim().ToLower()) || 
                                 u.Name.Trim().ToLower().Contains(s.Trim().ToLower()) ||
-                                u.Patronymic.Trim().ToLower().Contains(s.Trim().ToLower())));
+                                u.Patronymic.Trim().ToLower().Contains(s.Trim().ToLower()) ||
+                                u.Id.ToString() == s));
                 break;
             case 1:
                 _roles = Roles = new ObservableCollection<Role>(RoleImplementation.GetAllRoles().Result.Where(r => r.RoleName.Trim().ToLower().Contains(s.Trim().ToLower())));
